requirements.txt
# Core web framework
streamlit>=1.28.0

# HTTP requests and data processing
requests>=2.31.0
pandas>=2.0.3
numpy>=1.26.0

# Machine learning and NLP (optimized for deployment)
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.1.0,<2.2.0
google-generativeai>=0.3.2

# PDF processing
PyPDF2>=3.0.1

# Vector storage for embeddings
faiss-cpu>=1.7.4

# Additional dependencies for text processing
beautifulsoup4>=4.13.4
lxml>=6.0.0
rouge-score>=0.1.2

# Keyword extraction and text analysis
keybert>=0.9.0

# Data visualization
plotly>=6.2.0

# Natural language processing utilities
nltk>=3.9.1

# OCR capabilities (optional)
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.1.0

# Additional dependencies for embeddings
huggingface-hub>=0.17.3

# Environment management
python-dotenv>=1.0.0

# Security and performance
watchdog>=3.0.0